backend/requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
python-dotenv
psycopg2-binary
bcrypt
PyJWT